flask/requirements.txt
Flask
flask_cors
Flask-API
PyYAML
mysql-connector
elasticsearch
kafka-python
Flask-Sockets
Flask-SocketIO
flask_api
boto3